+++ engine/geometry/IModel.h
@@ -9,5 +9,5 @@ public:
 	virtual void accept(IRenderer* renderer) const = 0;
 
 	virtual Box6<coord> bounds() const = 0;
-	virtual int polygons() const = 0;
+	virtual size_t polygons() const = 0;
 };
